+++ src/edu/sjsu/stratagem/Environment.java
@@ -1,5 +1,7 @@
 package edu.sjsu.stratagem;
 
+import edu.sjsu.stratagem.exception.StratagemException;
+
 import java.util.Map;
 import java.util.HashMap;
