+++ src/edu/sjsu/stratagem/ExpressionBuilderVisitor.java
@@ -3,6 +3,7 @@ package edu.sjsu.stratagem;
 import java.util.ArrayList;
 import java.util.List;
 
+import edu.sjsu.stratagem.exception.StratagemException;
 import edu.sjsu.stratagem.parser.StratagemBaseVisitor;
 import edu.sjsu.stratagem.parser.StratagemParser;
